requirements.txt
flask==2.1.0
openai
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.2.1
python-dotenv==0.21.0
serpapi
